[#1] Fixed Gitian Builds, changed addresses starting letter and added a proper spork key
Gitian now has proper links for building, a skeleton was left for a later implementation of signing.

Addresses will now start with either c or 9 as in testnet privkeys. This is the quickest workaround to get it changed.

A proper spork key was added and given to who i was contacted by. I still hold a copy myself. Sporks are now enabled